+++ inference_multibeam.py
@@ -35,10 +35,10 @@ def read_input(readers, t0, a=None, tstep=1024, nchan=320):
     nbeams = len(readers)
     u8 = (readers[0].header['nbits'] == 8)
     if a is None:
-        a = np.zeros((nbeams, tstep, nchan, 1), dtype=np.float32)
+        a = np.zeros((nbeams, tstep, nchan, 1), dtype=np.uint8)
     for i in range(nbeams):
         readers[i].read_data(t_start=t0, t_stop=t0+tstep)
-        a[i, ..., 0] = readers[i].data.squeeze().astype('uint8').astype(np.float32)
+        a[i, ..., 0] = readers[i].data.squeeze().astype('uint8')
     return a
 
 def filter_detection(detections, n=3):
